@@ -2,37 +2,49 @@
   "name": "gadgetify",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "@prisma/client": "^5.8.1",
-    "cloudinary": "^1.41.1",
-    "jsonwebtoken": "^9.0.2",
-    "next": "13.4.18",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-icons": "^4.12.0",
-    "react-toastify": "^9.1.3",
-    "sharp": "^0.33.1",
-    "swiper": "^11.0.5",
-    "zod": "^3.22.4"
+    "@prisma/adapter-pg": "^7.4.0",
+    "@prisma/client": "^7.4.0",
+    "@tailwindcss/postcss": "^4.1.18",
+    "@types/bcryptjs": "^3.0.0",
+    "bcryptjs": "^3.0.3",
+    "dotenv": "^17.3.1",
+    "jsonwebtoken": "^9.0.3",
+    "next": "16.1.6",
+    "pg": "^8.18.0",
+    "react": "19.2.4",
+    "react-dom": "19.2.4",
+    "react-icons": "^5.5.0",
+    "react-toastify": "^11.0.5",
+    "sharp": "^0.34.5",
+    "swiper": "^12.1.1",
+    "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@types/jsonwebtoken": "^9.0.5",
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.0.1",
-    "daisyui": "^4.4.24",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.4",
-    "postcss": "^8",
-    "prisma": "^5.8.1",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5"
-  }
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/node": "^25.2.3",
+    "@types/pg": "^8.16.0",
+    "@types/react": "19.2.14",
+    "@types/react-dom": "19.2.3",
+    "autoprefixer": "^10.4.24",
+    "daisyui": "^5.5.18",
+    "eslint": "^9.20.1",
+    "eslint-config-next": "16.1.6",
+    "postcss": "^8.5.6",
+    "prisma": "^7.4.0",
+    "tailwindcss": "^4.1.18",
+    "typescript": "^5.9.3"
+  },
+  "overrides": {
+    "@types/react": "19.2.14",
+    "@types/react-dom": "19.2.3"
+  },
+  "browserslist": "> 1%"
 }
