@@ -1,29 +1,35 @@
 import { NextRequest, NextResponse } from "next/server";
-import prisma from "@/prisma/client";
-import ApiError from "../../utils/ApiError";
-import ApiResponse from "../../utils/ApiResponse";
-import { deleteFiles, fileToUrl } from "../../utils/files";
-import { userSchema } from "../route";
+import { prisma } from "@/lib/prisma";
+import ApiError from "@/app/api/utils/ApiError";
+import ApiResponse from "@/app/api/utils/ApiResponse";
+import { z } from "zod";
 import bcryptjs from "bcryptjs";
-import jwt from "jsonwebtoken";
-import generateToken from "../../utils/GenerateToken";
+
+const userSchema = z.object({
+  name: z.string().optional(),
+  phone: z.string().startsWith("01").length(11),
+  password: z.string().optional(),
+  address: z.string(),
+});
 
 interface Props {
-  params: { id: string };
+  params: Promise<{ id: string }>;
 }
 
 export async function GET(req: NextRequest, { params }: Props) {
+  const { id } = await params;
   let user;
   try {
     user = await prisma.user.findUnique({
-      where: { id: Number(params.id) },
+      where: { id },
     });
   } catch (error: any) {
     return NextResponse.json(
-      new ApiError(401, error.message || "Can't find details")
+      new ApiError(401, error.message || "Can't find details"),
+      { status: 401 }
     );
   }
-  if (!user || user === null)
+  if (!user)
     return NextResponse.json(new ApiError(404, "User not found"), {
       status: 404,
     });
@@ -31,36 +37,28 @@ export async function GET(req: NextRequest, { params }: Props) {
   return NextResponse.json(
     new ApiResponse(200, {
       id: user.id,
-      fullname: user.fullname,
-      rollNo: user.rollNo,
-      session: user.session,
-      year: user.year,
+      name: user.name,
       phone: user.phone,
-      email: user.email,
-      interests: user.interests,
-      images: user.images,
-      isVerified: user.isVerified,
-      isAdmin: user.isAdmin,
-      membershipFee: user.membershipFee,
-      membershipValidity: user.membershipValidity,
-      membershipType: user.membershipType,
-      memberId: user.memberId,
+      address: user.address,
+      createdAt: user.createdAt,
     }),
     { status: 200 }
   );
 }
 
 export async function PUT(req: NextRequest, { params }: Props) {
+  const { id } = await params;
   const data = await req.formData();
-  const file: any = data.get("images");
 
   const prevData = await prisma.user.findUnique({
-    where: {
-      id: parseInt(params.id),
-    },
+    where: { id },
   });
 
-  if (prevData?.phone !== String(data.get("phone") || "")) {
+  if (!prevData) {
+    return NextResponse.json(new ApiError(404, "User not found"), { status: 404 });
+  }
+
+  if (prevData.phone !== String(data.get("phone") || "")) {
     let user = await prisma.user.findFirst({
       where: {
         phone: String(data.get("phone") || ""),
@@ -73,162 +71,71 @@ export async function PUT(req: NextRequest, { params }: Props) {
       );
     }
   }
-  if (prevData?.rollNo !== String(data.get("rollNo") || "")) {
-    let user = await prisma.user.findFirst({
-      where: {
-        rollNo: String(data.get("rollNo") || ""),
-      },
-    });
-    if (user) {
-      return NextResponse.json(
-        new ApiError(400, "User already exist with this Roll number"),
-        { status: 400 }
-      );
-    }
-  }
 
-  let image: string = prevData?.images!;
-
-  if (file.size > 1) {
-    try {
-      await deleteFiles(image, "users"); // deleting the previous files
-    } catch (e: any) {
-      return NextResponse.json(
-        new ApiError(
-          420,
-          e.message || "There have a problem to delete previous images"
-        ),
-        { status: 420 }
-      );
-    }
-    try {
-      image = await fileToUrl(file, "users");
-    } catch (e: any) {
-      return NextResponse.json(
-        new ApiError(420, e.message || "There have a problem to upload avatar"),
-        { status: 420 }
-      );
-    }
-  }
   let body: any = {
-    fullname: data.get("fullname"),
-    rollNo: data.get("rollNo") || "",
-    session: data.get("session"),
-    year: data.get("year") || "",
+    name: data.get("name"),
     phone: data.get("phone"),
-    email: data.get("email") || "",
-    interests: data.get("interests"),
+    address: data.get("address"),
     password: data.get("password") || "",
-    images: image,
   };
-  if (!body.phone)
-    return NextResponse.json(new ApiError(404, "Phone number can't be blank"), {
-      status: 404,
-    });
-  else if (body.rollNo.length !== 10) {
-    return NextResponse.json(
-      new ApiError(402, "Roll Number must have 10 digits"),
-      { status: 402 }
-    );
-  } else if (body.phone.length !== 11) {
-    return NextResponse.json(
-      new ApiError(402, "Phone number must be 11 digits"),
-      { status: 402 }
-    );
-  } else if (body.phone[0] !== "0" || body.phone[1] !== "1") {
-    return NextResponse.json(
-      new ApiError(402, "Phone number must be start with 01"),
-      { status: 402 }
-    );
-  }
 
-  const validatedData: any = userSchema.safeParse(body);
-  if (!validatedData.success) {
+  const validation = userSchema.safeParse(body);
+  if (!validation.success) {
     return NextResponse.json(
       new ApiError(
         400,
-        validatedData.error.errors[0].message || "Invalid Input",
-        validatedData.error
+        validation.error.issues[0].message || "Invalid Input",
+        validation.error.issues
       ),
-      {
-        status: 400,
-      }
+      { status: 400 }
     );
   }
 
-  if (body.password === "") {
-    delete validatedData.data.password;
-  } else {
+  const updateData: any = {
+    name: validation.data.name,
+    phone: validation.data.phone,
+    address: validation.data.address,
+  };
+
+  if (validation.data.password) {
     const salt = await bcryptjs.genSalt(10);
-    validatedData.data.password = await bcryptjs.hash(
-      validatedData.data.password,
-      salt
-    );
+    updateData.password = await bcryptjs.hash(validation.data.password, salt);
   }
 
   let updatedUser;
   try {
     updatedUser = await prisma.user.update({
-      where: { id: Number(params.id) },
-      data: validatedData.data,
+      where: { id },
+      data: updateData,
     });
   } catch (e: any) {
     return NextResponse.json(
-      new ApiError(404, e.message || "User data did not update"),
-      { status: 404 }
+      new ApiError(500, e.message || "User data did not update"),
+      { status: 500 }
     );
-  }
-
-  //create token data
-  const tokenData = {
-    id: updatedUser.id,
-    images: updatedUser.images,
-    isAdmin: updatedUser.isAdmin,
-  };
-
-  const token = generateToken(tokenData);
-
-  const res = NextResponse.json(
-    new ApiResponse(202, updatedUser, "User details updated Successfully"),
-    {
-      status: 202,
-    }
-  );
-
-  res.cookies.set("token", token, {
-    httpOnly: true,
-  });
-  return res;
-}
-
-export async function DELETE(req: NextRequest, { params }: Props) {
-  const prevData = await prisma.user.findFirst({
-    where: {
-      id: parseInt(params.id),
-    },
-  });
-
-  let image: string = prevData?.images!;
-  try {
-    await deleteFiles(image, "users");
-  } catch (error: any) {
-    return NextResponse.json(
-      new ApiError(420, error.message || "Can't delete images")
-    );
-  } // deleting the previous files
-
-  const user = await prisma.user.delete({
-    where: { id: Number(params.id) },
-  });
-
-  if (!user || !prevData) {
-    return NextResponse.json(new ApiError(400, "User can't be deleted"), {
-      status: 400,
-    });
   }
 
   return NextResponse.json(
-    new ApiResponse(202, "", "User deleted successfully"),
+    new ApiResponse(202, updatedUser, "User details updated Successfully"),
     { status: 202 }
   );
+}
+
+export async function DELETE(req: NextRequest, { params }: Props) {
+  const { id } = await params;
+
+  try {
+    const user = await prisma.user.delete({
+      where: { id },
+    });
+    return NextResponse.json(
+      new ApiResponse(202, "", "User deleted successfully"),
+      { status: 202 }
+    );
+  } catch (error: any) {
+    return NextResponse.json(
+      new ApiError(400, error.message || "User can't be deleted"),
+      { status: 400 }
+    );
+  }
 }
