@@ -1,11 +1,36 @@
 import Hero from "./components/Hero";
 import Products from "./components/Products";
+import Features from "./components/Features";
+import Stats from "./components/Stats";
+import Newsletter from "./components/Newsletter";
+import { prisma } from "@/lib/prisma";
 
-export default function Home() {
+export const metadata = {
+  title: "Home | Gadgetify",
+  description: "Browse our exclusive collection of latest gadgets and tech accessories.",
+};
+
+export default async function Home() {
+  let products: any[] = [];
+  try {
+    // Limit products on home page to latest 12 for better performance
+    products = await prisma.product.findMany({
+      take: 20,
+      orderBy: {
+        createdAt: 'desc'
+      }
+    });
+  } catch (error) {
+    console.error("Error fetching products:", error);
+  }
+
   return (
-    <main>
+    <main className="overflow-hidden">
       <Hero />
-      <Products />
+      <Features />
+      <Products initialProducts={products} />
+      <Stats />
+      <Newsletter />
     </main>
   );
 }
