@@ -1,28 +1,38 @@
 import type { Metadata } from "next";
 import SingleProduct from "./SingleProduct";
 
-let id: string;
-const Page = ({ params }: { params: { id: string; slug: string } }) => {
-  id = params.id;
-  return <SingleProduct params={params} />;
+const Page = async ({ params }: { params: Promise<{ id: string; slug: string }> }) => {
+  const awaitedParams = await params;
+  const data = await fetch(
+    `${process.env.NEXT_PUBLIC_BASE_URL}/api/product/${awaitedParams.id}`,
+    { cache: "no-store" }
+  );
+  const jData = await data.json();
+  const product = jData.success ? jData.data : null;
+
+  return <SingleProduct params={awaitedParams} initialProduct={product} />;
 };
 
-export async function generateMetadata(): Promise<Metadata> {
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ id: string; slug: string }>;
+}): Promise<Metadata> {
+  const awaitedParams = await params;
   const data = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/api/product/${id}`
+    `${process.env.NEXT_PUBLIC_BASE_URL}/api/product/${awaitedParams.id}`
   );
   const jData = await data.json();
   if (jData.success) {
-    const product = await jData.data;
+    const product = jData.data;
     return {
       title: `${product.name} - Gadgetify`,
-      description: `${product.description} || Gadgetify is a one-stop shop for all your gadgets needs. We sell everything from phones to laptops to smartwatches and more!  Come visit us today!`,
+      description: `${product.description} || Gadgetify is a one-stop shop for all your gadgets needs.`,
     };
   } else {
     return {
       title: "Gadgetify",
-      description:
-        "Gadgetify is a one-stop shop for all your gadgets needs. We sell everything from phones to laptops to smartwatches and more!  Come visit us today!",
+      description: "Gadgetify is a one-stop shop for all your gadgets needs.",
     };
   }
 }
